requirements.txt
pinecone
langchain
pypdf2
python-dotenv
openai
fastapi  
uvicorn
sentence_transformers>=3.0.0
transformers<=4.50.0
pdfplumber
PyMuPDF